requirements.txt
dashscope>=1.10.0
pdf2image>=1.16.3
Pillow>=10.0.0
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0